// vava-pentecoste-system/src/app/models/item.ts
export interface ItemInterface {
  $key?: string;
  available?: string;
  category?: string;
  description?: string;
  product_id?: string;
  image?: string;
  name?: string;
  price?: number;
  stock?: number;
  categories?: string;
  percent?: string;
  popularity?: number;
  image_firebase_url?: string;
  lenght?: string;
}